♻️ refactor(logs): move logs page into feature module
Move the logs page implementation, styles, parsing helpers, filters, and scroller hooks into src/features/logs.

Keep src/pages/LogsPage.tsx as a route wrapper and remove the remaining logs-specific code from pages.
