+++ src/features/logs/LogsPage.module.scss
@@ -1,4 +1,4 @@
-@use '../styles/mixins' as *;
+@use '../../styles/mixins' as *;
 
 .container {
   width: 100%;
